Refactored Tree view to render Tree component
